output revert.
