+++ CK2ToEU4/CK2ToEU4.vcxproj
@@ -208,6 +208,7 @@
     <ClCompile Include="Source\EU4World\Output\outModFile.cpp" />
     <ClCompile Include="Source\EU4World\Output\outProvince.cpp" />
     <ClCompile Include="Source\EU4World\Output\outVersion.cpp" />
+    <ClCompile Include="Source\EU4World\Output\outWorld.cpp" />
     <ClCompile Include="Source\EU4World\Province\EU4Province.cpp" />
     <ClCompile Include="Source\EU4World\Province\ProvinceDetails.cpp" />
     <ClCompile Include="Source\EU4World\Province\ProvinceModifier.cpp" />
@@ -254,8 +255,10 @@
     <ClInclude Include="Source\EU4World\Country\Country.h" />
     <ClInclude Include="Source\EU4World\Country\CountryDetails.h" />
     <ClInclude Include="Source\EU4World\EU4World.h" />
-    <ClInclude Include="Source\EU4World\Output\ModFile.h" />
-    <ClInclude Include="Source\EU4World\Output\output.h" />
+    <ClInclude Include="Source\EU4World\Output\outProvince.h" />
+    <ClInclude Include="Source\EU4World\Output\outCountry.h" />
+    <ClInclude Include="Source\EU4World\Output\outModFile.h" />
+    <ClInclude Include="Source\EU4World\Output\outVersion.h" />
     <ClInclude Include="Source\EU4World\Province\EU4Province.h" />
     <ClInclude Include="Source\EU4World\Province\ProvinceDetails.h" />
     <ClInclude Include="Source\EU4World\Province\ProvinceModifier.h" />
