+++ CK2ToEU4/Source/EU4World/EU4World.cpp
@@ -1,16 +1,12 @@
 #include "EU4World.h"
-#include "../CK2World/Characters/Character.h"
-#include "../CK2World/Titles/Title.h"
-#include "../CK2World/World.h"
-#include "../Configuration/Configuration.h"
-#include "../Mappers/VersionParser/VersionParser.h"
-#include "Country/Country.h"
 #include "Log.h"
 #include "OSCompatibilityLayer.h"
-#include "Province/EU4Province.h"
 #include <filesystem>
 #include <fstream>
 namespace fs = std::filesystem;
+#include "../CK2World/Characters/Character.h"
+#include "../CK2World/Titles/Title.h"
+#include "../Configuration/Configuration.h"
 
 EU4::World::World(const CK2::World& sourceWorld, const Configuration& theConfiguration, const mappers::VersionParser& versionParser)
 {
@@ -138,100 +134,3 @@ void EU4::World::importVanillaCountries(const std::string& eu4Path)
 	LOG(LogLevel::Info) << ">> Loaded " << fileNames.size() << " history files.";
 }
 
-void EU4::World::output(const mappers::VersionParser& versionParser, const Configuration& theConfiguration) const
-{
-	LOG(LogLevel::Info) << "<- Creating Output Folder";
-	fs::create_directory("output");
-	if (fs::exists("output/" + theConfiguration.getOutputName())) {
-		Log(LogLevel::Info) << "<< Deleting existing mod filder.";
-		fs::remove_all("output/" + theConfiguration.getOutputName());
-	}
-
-	fs::create_directory("output/" + theConfiguration.getOutputName());
-	fs::create_directory("output/" + theConfiguration.getOutputName() + "/history/");
-	fs::create_directory("output/" + theConfiguration.getOutputName() + "/history/countries/");
-	fs::create_directory("output/" + theConfiguration.getOutputName() + "/history/provinces/");
-	fs::create_directory("output/" + theConfiguration.getOutputName() + "/common/");
-	fs::create_directory("output/" + theConfiguration.getOutputName() + "/common/countries/");
-	fs::create_directory("output/" + theConfiguration.getOutputName() + "/common/country_tags/");
-
-	LOG(LogLevel::Info) << "<- Crafting .mod File";
-	createModFile(theConfiguration);
-
-	// Record converter version
-	LOG(LogLevel::Info) << "<- Writing version";
-	outputVersion(versionParser, theConfiguration);
-
-	// Output common\countries.txt
-	LOG(LogLevel::Info) << "<- Creating countries.txt";
-	outputCommonCountriesFile(theConfiguration);
-
-	LOG(LogLevel::Info) << "<- Writing Countries";
-	outputCommonCountries(theConfiguration);
-	outputHistoryCountries(theConfiguration);
-
-	LOG(LogLevel::Info) << "<- Writing Provinces";
-	outputHistoryProvinces(theConfiguration);
-}
-
-void EU4::World::createModFile(const Configuration& theConfiguration) const
-{
-	std::ofstream output("output/" + theConfiguration.getOutputName() + ".mod");
-	if (!output.is_open()) throw std::runtime_error("Could not create " + theConfiguration.getOutputName() + ".mod");
-	LOG(LogLevel::Info) << "<< Writing to: "
-							  << "output/" + theConfiguration.getOutputName() + ".mod";
-	output << modFile;
-	output.close();
-}
-
-
-void EU4::World::outputVersion(const mappers::VersionParser& versionParser, const Configuration& theConfiguration) const
-{
-	std::ofstream output("output/" + theConfiguration.getOutputName() + "/ck2toeu4_version.txt");
-	if (!output.is_open()) throw std::runtime_error("Error writing version file! Is the output folder writable?");
-	output << versionParser;
-	output.close();
-}
-
-void EU4::World::outputCommonCountriesFile(const Configuration& theConfiguration) const
-{
-	std::ofstream output("output/" + theConfiguration.getOutputName() + "/common/country_tags/00_countries.txt");
-	if (!output.is_open()) throw std::runtime_error("Could not create countries file!");
-
-	for (const auto& country: countries) { output << country.first << " = \"" << country.second->getCommonCountryFile() << "\"\n"; }
-	output << "\n";
-	output.close();
-}
-
-void EU4::World::outputHistoryProvinces(const Configuration& theConfiguration) const
-{
-	for (const auto& province: provinces) {
-		std::ofstream output("output/" + theConfiguration.getOutputName() + "/" + province.second->getHistoryCountryFile());
-		if (!output.is_open())
-			throw std::runtime_error("Could not create country history file: output/" + theConfiguration.getOutputName() + "/" + province.second->getHistoryCountryFile());
-		output << *province.second;
-		output.close();
-	}
-}
-
-void EU4::World::outputHistoryCountries(const Configuration& theConfiguration) const
-{
-	for (const auto& country: countries) {
-		std::ofstream output("output/" + theConfiguration.getOutputName() + "/" + country.second->getHistoryCountryFile());
-		if (!output.is_open())
-			throw std::runtime_error("Could not create country history file: output/" + theConfiguration.getOutputName() + "/" + country.second->getHistoryCountryFile());
-		output << *country.second;
-		output.close();
-	}
-}
-
-void EU4::World::outputCommonCountries(const Configuration& theConfiguration) const
-{
-	for (const auto& country: countries) {
-		std::ofstream output("output/" + theConfiguration.getOutputName() + "/" + country.second->getCommonCountryFile());
-		if (!output.is_open())
-			throw std::runtime_error("Could not create country common file: output/" + theConfiguration.getOutputName() + "/" + country.second->getCommonCountryFile());
-		country.second->outputCommons(output);
-		output.close();
-	}
-}
