+++ CK2ToEU4/Source/EU4World/Output/outCountry.h
@@ -1,0 +1,11 @@
+#ifndef OUT_COUNTRY_H
+#define OUT_COUNTRY_H
+
+#include "../Country/Country.h"
+#include <ostream>
+
+namespace EU4
+{
+std::ostream& operator<<(std::ostream& output, const Country& country);
+}
+#endif // OUT_COUNTRY_H
