+++ CK2ToEU4/Source/EU4World/Output/outProvince.h
@@ -1,0 +1,13 @@
+#ifndef OUT_PROVINCE_H
+#define OUT_PROVINCE_H
+
+#include "../Province/EU4Province.h"
+#include <ostream>
+
+namespace EU4
+{
+std::ostream& operator<<(std::ostream& output, const Province& province);
+} // namespace EU4
+
+
+#endif // OUT_PROVINCE_H
