--- CK2ToEU4/Source/EU4World/Output/output.h
@@ -1,22 +1,0 @@
-#ifndef OUTPUT_H
-#define OUTPUT_H
-
-#include "../../Mappers/VersionParser/VersionParser.h"
-#include "../Country/Country.h"
-#include "../Province/EU4Province.h"
-#include "ModFile.h"
-#include <ostream>
-
-namespace EU4
-{
-std::ostream& operator<<(std::ostream& output, const Country& country);
-std::ostream& operator<<(std::ostream& output, const ModFile& modFile);
-std::ostream& operator<<(std::ostream& output, const Province& province);
-} // namespace EU4
-
-namespace mappers
-{
-std::ostream& operator<<(std::ostream& output, const VersionParser& versionParser);
-} // namespace mappers
-
-#endif // OUTPUT_H
